Bugfix for unicode paths
