+++ athentoimaging/lines_detection_p.py
@@ -1,6 +1,7 @@
 import cv2 as cv
 import numpy as np
 import os
+import img_utils as iu
 
 
 def detect_lines(input_file,
@@ -10,7 +11,7 @@ def detect_lines(input_file,
 
     if min_val < max_val:
 
-        image = get_image(input_file)
+        image = iu.get_image(input_file)
 
         check_canny_args(min_val, max_val, aperture_size)
         check_houghlines_p_args(rho, theta, threshold,
@@ -28,7 +29,7 @@ def detect_lines(input_file,
 
 def delete_lines(input_file, lines, width=5, color=(255, 255, 255)):
 
-    image = get_image(input_file)
+    image = iu.get_image(input_file)
 
     check_lines(lines)
     check_width(width)
@@ -43,7 +44,7 @@ def delete_all_lines(input_file,
                      rho=1, theta=np.pi/180, threshold=200,
                      min_line_length=30, max_line_gap=20):
 
-    image = get_image(input_file)
+    image = iu.get_image(input_file)
 
     check_width(width)
     check_color(color)
@@ -95,7 +96,7 @@ def distance_mean(lines):
 
 def draw_lines(input_file, lines, width=5, color=(0, 0, 255)):
 
-    image = get_image(input_file)
+    image = iu.get_image(input_file)
 
     check_lines(lines)
     check_width(width)
@@ -155,20 +156,6 @@ def parallels(line1, line2, error=5):
 
 
 #Checking arguments
-
-def get_image(input_file):
-    if input_file == '' or input_file is None:
-            raise ValueError("Input_file must be different than '' or None.")
-
-    image = input_file
-
-    if isinstance(input_file, str):
-        if os.path.isfile(input_file) is False:
-            raise IOError("Input file not found.")
-        image = cv.imread(input_file)
-
-    return image
-
 
 def check_color(color):
